More work on edge looping
Worked some more on getting the tiles to loop when auto arranging when
they hit the edges of the screen.  It's mostly working, but I'm working
on cases where it's next to two edges and trying to decide which way to
go.  I think that is even working, except perhaps in cases where the
width of a starting row is wider than the window itself.